inlined implementation to header files and removed superfluous definitions
N.B. the aim is to make the disruptor also a header-only library that can be templated and more efficiently inlined/constexpr-optimised by the compiler

+++ src/disruptor/include/disruptor/ITaskScheduler.hpp
@@ -8,7 +8,7 @@ class ITaskScheduler {
 public:
     virtual ~ITaskScheduler()                                                     = default;
 
-    virtual void              start(std::int32_t numberOfThreads)                 = 0;
+    virtual void              start(std::size_t numberOfThreads)                  = 0;
     virtual void              stop()                                              = 0;
 
     virtual std::future<void> scheduleAndStart(std::packaged_task<void()> &&task) = 0;
